Rename package from pieces to piece
