+++ src/main/java/piece/Bishop.java
@@ -1,6 +1,6 @@
-package pieces;
+package piece;
 
-import pieces.attribute.*;
+import piece.attribute.*;
 
 public class Bishop extends Piece {
     Bishop(Color color) {
