+++ src/main/java/piece/Blank.java
@@ -1,6 +1,6 @@
-package pieces;
+package piece;
 
-import pieces.attribute.*;
+import piece.attribute.*;
 
 public class Blank extends Piece {
     Blank(Color color) {
