+++ src/main/java/piece/Queen.java
@@ -1,6 +1,6 @@
-package pieces;
+package piece;
 
-import pieces.attribute.*;
+import piece.attribute.*;
 
 public class Queen extends Piece {
     Queen(Color color) {
