+++ src/main/java/piece/attribute/Color.java
@@ -1,4 +1,4 @@
-package pieces.attribute;
+package piece.attribute;
 
 public enum Color {
     WHITE, BLACK, NO_COLOR;
